pyproject.toml
[project]
name = "nba-duckdb-dbt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb>=1.9.3",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "polars>=1.28.1",
    "pyarrow>=20.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]